+++ ChangeLog.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 ### Enhancements
 
 - Bumped the OpenAI Java library to v4.37.0
+- [#11](https://github.com/jlangch/venice/issues/11): Migrated to JLine 3.25.1. 
+  JLine 3.25.1 includes JAnsi